Move webserver into it's own folder
